Split git-contrib-graph file into packages
